\docType{import}
\name{reexports}
\alias{reexports}
\alias{tidy}
\alias{glance}
\title{Objects exported from other packages}
\keyword{internal}
\description{
  These objects are imported from the "generics" package.
  See \code{\link[generics:tidy]{tidy}} and
  \code{\link[generics:glance]{glance}} for details.
}
